feat: implement PageRollupProperty and PageButtonProperty structures with serialization support
