+++ notionrs/src/client/page/get_page_property_item.rs
@@ -10,9 +10,7 @@ pub struct GetPagePropertyItemClient {
 
 impl GetPagePropertyItemClient {
     // TODO: docs for send
-    pub async fn send(
-        self,
-    ) -> Result<crate::object::page::properties::PageProperty, crate::error::Error> {
+    pub async fn send(self) -> Result<crate::object::page::PageProperty, crate::error::Error> {
         let page_id = self.page_id.ok_or(crate::error::Error::RequestParameter(
             "`page_id` is not set.".to_string(),
         ))?;
@@ -44,7 +42,7 @@ impl GetPagePropertyItemClient {
             .await
             .map_err(|e| crate::error::Error::BodyParse(e.to_string()))?;
 
-        let user = serde_json::from_slice::<crate::object::page::properties::PageProperty>(&body)?;
+        let user = serde_json::from_slice::<crate::object::page::PageProperty>(&body)?;
 
         Ok(user)
     }
