+++ notionrs/src/client/page/update_page.rs
@@ -1,7 +1,5 @@
 use serde::{Deserialize, Serialize};
 
-use crate::object::page::page_response::PageResponse;
-
 #[derive(Debug, Default, notionrs_macro::Setter)]
 pub struct UpdatePageClient {
     /// The reqwest http client
@@ -9,8 +7,7 @@ pub struct UpdatePageClient {
 
     pub(crate) page_id: Option<String>,
 
-    pub(crate) properties:
-        std::collections::HashMap<String, crate::object::page::properties::PageProperty>,
+    pub(crate) properties: std::collections::HashMap<String, crate::object::page::PageProperty>,
 
     pub(crate) icon: Option<crate::object::icon::Icon>,
 
@@ -19,8 +16,7 @@ pub struct UpdatePageClient {
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct UpdatePageRequestBody {
-    pub(crate) properties:
-        std::collections::HashMap<String, crate::object::page::properties::PageProperty>,
+    pub(crate) properties: std::collections::HashMap<String, crate::object::page::PageProperty>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     pub(crate) icon: Option<crate::object::icon::Icon>,
@@ -30,7 +26,7 @@ pub struct UpdatePageRequestBody {
 }
 
 impl UpdatePageClient {
-    pub async fn send(self) -> Result<PageResponse, crate::error::Error> {
+    pub async fn send(self) -> Result<crate::object::page::PageResponse, crate::error::Error> {
         let page_id = self.page_id.ok_or(crate::error::Error::RequestParameter(
             "`page_id` is not set.".to_string(),
         ))?;
@@ -65,7 +61,7 @@ impl UpdatePageClient {
             .await
             .map_err(|e| crate::error::Error::BodyParse(e.to_string()))?;
 
-        let page: PageResponse = serde_json::from_slice::<PageResponse>(&body)?;
+        let page = serde_json::from_slice::<crate::object::page::PageResponse>(&body)?;
 
         Ok(page)
     }
